backend/requirements.txt
mpu6050-raspberrypi>=1.2
smbus2>=0.4.2
psutil>=5.9.0